routes gefixed dat / voor home is en //{id} voor details.
article-list.blade.php en search-filter.blade.php zijn nu partials die je terug ziet in home.blade.php

+++ app/Http/Controllers/ArticleController.php
@@ -5,17 +5,30 @@ namespace App\Http\Controllers;
 use App\Models\Age;
 use App\Models\Article;
 use App\Models\Subject;
+use App\Models\User;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
+use Spatie\Permission\Models\Role;
 
 class ArticleController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index(Request $request)
+    public function index(Request $request, User $user)
     {
+        $adminUser = User::find(1);
+        $users = User::all();
+        $adminRole = Role::findByName('admin');
+        $moderatorRole = Role::findByName('moderator');
+        $userRole = Role::findByName('user');
+
+        $adminUser->assignRole($adminRole, $userRole, $moderatorRole);
+        foreach ($users as $user) {
+            $user->assignRole($userRole);
+        }
 //        filters en search results ophalen
         $selectedSubjects = $request->input('subjects', []);
         $selectedAges = $request->input('ages', []);
@@ -54,7 +67,7 @@ class ArticleController extends Controller
         $ages = Age::all();
         $subjects = Subject::all();
 
-        return view('home', compact('articles', 'ages', 'subjects','selectedAges', 'selectedSubjects', 'searchInput'));
+        return view('home', compact('articles', 'ages', 'subjects','selectedAges', 'selectedSubjects', 'searchInput', 'user'));
     }
 
     /**
@@ -106,9 +119,11 @@ class ArticleController extends Controller
 
     /**
      * Show the form for creating a new resource.
+     * @throws AuthorizationException
      */
-    public function create()
+    public function create(Article $article)
     {
+        $this->authorize('create', $article);
         $ages = Age::all();
         $subjects = Subject::all();
 
@@ -159,9 +174,11 @@ class ArticleController extends Controller
 
     /**
      * Show the form for editing the specified resource.
+     * @throws AuthorizationException
      */
-    public function edit(string $id)
+    public function edit(string $id, Article $article)
     {
+        $this->authorize('edit', $article);
         $article = Article::find($id);
         $ages = Age::all();
         $subjects = Subject::all();
@@ -213,9 +230,11 @@ class ArticleController extends Controller
 
     /**
      * Remove the specified resource from storage.
+     * @throws AuthorizationException
      */
-    public function destroy(string $id)
+    public function destroy(string $id, Article $article)
     {
+        $this->authorize('delete', $article);
         $article = Article::find($id);
 
         $article->delete();
